fix(Patient Portal): convert appointment list as a component
